Fix tessellation sample resolution calculation
Remove the "optimization" that tried to avoid calculating the edge
length by simply subdividing the parent edge's length -- that doesn't
work correctly when sub-edges are significantly off parallel.  The
sample resolution must be compared with the true edge length to judge
when to stop sampling.

Also add a simple mechanism to allow tessellation functions to specify
the sample resolution.

git-archimport-id: miles@gnu.org--gnu-2005/snogray--devo--0--patch-93