Improve memory allocation in Tessel class
Instead of using new/delete with an ad-hoc freelist, make a real
freelist class and use that.  This wastes less space, is able to return
memory to the OS, and reduces the need to keep track of all temporary
objects.  There seems to be somewhat better locality in the resulting
allocations too.

git-archimport-id: miles@gnu.org--gnu-2005/snogray--devo--0--patch-101