Fix memory corruption problem with freelists
When a call to the (very simple) freelist allocation function was
inlined into a constructor, gcc's aliasing rules made it think that it
was safe to re-order the constructor's initialization code before the
allocation code, because the manifest types of the newly allocated
object and the freelist's internal freelist node are different (when of
course they point to the same memory).

To protect against this, give the freelist's list-head pointer a
"may_alias" attribute.  [This is gcc-specific; is there any portable way
to do this?  Use char* somehow?]

git-archimport-id: miles@gnu.org--gnu-2005/snogray--devo--0--patch-105