Improve polygon tessellation in .nff file loader
The old method (generating a fan of triangles from one polygon vertex to
all the others) didn't work well with irregular polygons, so instead
make a center vertex for polygons with more than four vertices, and
generate a fan of triangles using that as one vertex (3- and 4-vertex
polygons are special cased to use the obvious tessellation).

With this change the SPD "gears" scene actually looks pretty good.

git-archimport-id: miles@gnu.org--gnu-2005/snogray--devo--0--patch-113