Remove reflected light from Mirror BRDF calculation
When calculating the contribution of a Mirror's BRDF, scale the result
by the inverse of the reflection amount.  The idea is that only
non-reflected light is available for the BRDF to operate on, but we
instead modify the BRDF output, as that's easier.

Also remove the scaling of the the phong highlight amount from the
Mirror constructor (which was a previous attempt to do something
similar, but provided less consistent results).

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-12