Rewrite illumination framework to gather samples before shadow-testing
Instead of doing shadow-testing for each light-ray individually, gather
light samples (pre-filtered through the BRDF), and then do shadow
testing on the entire set of samples.  This allows us to sort the
samples by contribution, and possibly avoid doing shadow-testing for
those which do not make a significant contribution to the result.

Hopefully this framework can also provide a better base for other
light-sampling optimizations.

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-19