Re-order arguments to phong BRDF; tweak test-scenes to preserve energy
The k_s parameter is now first, and no longer optional.

The BRDF parameter of the Material class is now a pointer instead of a
reference (a reference is nice because it prevents NULL values, but it
was just more awkward for callers; now a NULL brdf parameter results in
a lambertian BRDF).

The test-scenes now try to use more energy-preserving values of k_d and k_s.
This results in an overall dimming of specular materials, so the
brightness of most lights has also been increased to compensate.

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-27