Redo Glass material to be more physically accurate
It now does correct fresnel reflection/transmission (and shadowing to an
extent).

This change also removes most of the glass parameters -- it now has only
a Medium parameter -- as I'm not sure how most of them fit in with the
calculations (a medium is enough to do colored glass, at least).  It
also now does not use a brdf at all, so non-visible light sources will
not have any effect.  Maybe other parameters like reflectance and color,
and brdf interaction can be added back later.

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-79