Fix mirrored sphere intersection problems
The recent change drastically reducing the value of Eps caused the
sphere intersection method to start failing for reflections (because the
reflected ray would appear to intersect the reflecting sphere, at a very
short instance).  Use a hardwired constant value of (1e-7) instead of
Eps to reject self-intersection.

In the future it would be a good idea to use some more consistent
approach to the self-intersection problem.

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-81