Add per-triangle materials to Mesh, and remove PrimarySurface class
The basic reason for the PrimarySurface class was to allow Mesh
triangles not to store a Material pointer, so now that they do, it's
pointless.

[No extra space is needed:  by always having a fixed Material pointer,
we can get rid of the Surface::material() virtual method, and thus
looking up a surface's shadow_type via its material becomes fast enough
to eliminate the need for a per-surface cached value of shadow_type.
Eliminating the per-surface shadow_type field balances the space needed
by a Material pointer.]

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-105