Reduce variance in environmental lights with extreme constrast ranges
Previously, regions in an environmental light all had equal weight,
brighter areas being handled by making the regions there smaller.  This
works fine in many cases, but if the contrast range in an environment
map is very large, the smallest regions can't be made small enough to
accurately represent it (the smallest possible region is more or less a
single pixel).  The resulting mismatch between sampling probability and
brightness caused a very high variance when the light was sampled.

This change allows regions to have an integer weight, and they are
sampled with a probability equal to their weight.  This weight is used
so that very bright regions that can't be made any smaller will still be
sampled approximately in proportion to their brightness.

git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-302