Rewrite Surface::intersect interface
The new interface should work better with higher-order surfaces, and
avoids some bugs with the old interface.  It is also slightly slower,
because it doesn't avoid an intersection-test with a surface that
originates a ray; hopefully this can be optimized in the future (e.g. by
priming the negative-hit cache with the originating surface for planar
surfaces).

git-archimport-id: miles@gnu.org/snogray--devo--0--patch-48