Make Surface::intersect interface more abstract and powerful
Use a more abstract mechanism for passing information between the
initial call to Surface::intersect and the subsequent creation of an
Intersect object (which can be quite expensive, so we don't want to do
it for every surface intersection).

The basic mechanism is to return a Surface::IsecInfo object allocated
using the Mempool mechanism to avoid excessive allocation overhead.  Now
the Intersect object is created by a virtual method in subclasses of
Surface::IsecInfo (replacing the old Surface::intersect_info method).

This should make Surface::intersect flexible enough to handle arbitrary
higher-order surfaces (e.g. instances).

Also make Surface::shadows a virtual method so that various subclasses
can define optimized versions of it.

git-archimport-id: miles@gnu.org/snogray--devo--0--patch-52