Rewrite shadowing interfaces
Get rid of "shadow_caster" methods, which returned the shadowing object,
and replace them with simpler "shadow" methods which return the type of
shadow found.  This should be more instancing-friendly.

Add ShadowRay class, a subclass of Ray which holds extra info that may
be useful for shadow intersection testing.

Only record top-level objects as shadow-hints:  in the presence of
instances or other nested object spaces, if a nested object were
recorded as a shadow-hint, it could not be tested correctly without
somehow recording the intervening nesting structure as well, which is
too expensive/complex.

Don't immediately reject self-shadowing, as higher order surfaces may
shadow themselves.  Instead, allow a surface to specify when it's safe
to use such an optimization.  Reimplement self-shadowing rejection by
priming the negative-intersection cache, instead of directly doing such
a test.

git-archimport-id: miles@gnu.org/snogray--devo--0--patch-65