Make Space search interface more generic
This change removes some octree-specific details from the Space search
interface.  In particular, it moves the use of negative_isec_cache
(for mailboxing) and associated stat counters into the Octree class.

Those are only needed for pure space division schemes like an octree or
kdtree, and not for e.g. bounding-box based accelerators.