Make Space build interface more generic
Remove the simple Space:add method (which adds a single surface to a
space), and replace it with a mechanism where one first adds all
surfaces to a subclass of SpaceBuilder, and then call the
SpaceBuilder::make_space method to actually construct the space.

This allows more sophisticated algorithms that use global information
for building a space.

Also remove the Space::Stats structure and Space::stats method, as those
were really octree-specific.