Add support for a bundled Lua distribution
If configure finds a "lua" subdirectory, or if the user specifies
another directory using --with-lua-src, then the Makefile will
automatically build Lua there and link against that library.

This should make building slightly easier on systems without a Lua
library already installed.

The LUA_DIST variable may be used with "make dist" to add a "lua"
subdirectory to a distribution.