Use intersection normal frame for lighting calculations
Instead of doing lighting calculations in the world coordinate system,
use a frame of reference where the intersection normal has the fixed
value (0, 0, 1).  We call this the "normal frame".

Doing calculations in the normal frame is sometimes more efficient, and
using an explicit notion of a frame helps clarify the code.