Rename quadratic_isec to quadratic_roots, and tweak interface
The name "quadratic_roots" is slightly more descriptive, and there's
nothing about the function that specializes it for use in computing
surface intersection (though that's generally where it's used).

Also, return the number of roots found (0-2) instead of a boolean, and
return the roots themselves in an array instead of using multiple out
parameters.