Change handling of intensity scaling in ImageOutput
Get rid of the ImageOutput::exposure field, as the intensity_scale field
contains the same information.

Treat the "exposure" initialization parameter as a scale factor (instead
of as a "number of stops"), and use it to directly initialize
intensity_scale.

Conversion from number of stops to a scale factor for the -e
command-line option is now done during argument parsing.