Use proper ray coordinate system in Instance constructor
The previous removal of the Intersect::ray field, replacing it with the
Trace::ray field, broken instancing, because Trace::ray is always in the
world coordinate system, whereas the Intersect constructor expects it to
be in the local coordinate system (when that is different, as it is in
the case of instancing).

To fix this, pass in the ray in local coordinates to the Intersect
constructor; it isn't actually stored, just used for the initial
Intersect setup calculations.