configure.ac: Check more thoroughly for installed Lua
Lua installations tend to be a bit inconsistent about things like the
place where "lua.h" is placed, and what the lua library is named.

On debian, we can just use pkg-config to get the details, but most
systems don't use pkg-config for Lua.  If we can't find an installed Lua
header and library using pkg-config or in the default system locations,
try various common include directories and library names.