configure.ac: Remove obsolete workaround for PKG_CHECK_MODULES bug
PKG_CHECK_MODULES used to have a bug where it did not print any output
in the case where the package was not found, so configure.ac used to
explicitly print "no\n" in that case.

Recent versions of pkg-config seem to have fixed this bug, so remove
the workaround (systems with older versions of pkg-config will be a
bit uglier, but it's just a cosmetic problem anyway).