Switch to new sample-generation framework
Rewrite the Grid class to be a subclass of SampleGen instead of
Sample2Gen, change the Renderer class to use SampleSet, and modify the
top-level rendering driver accordingly.

This also changes the meaning of the "-a" command-line option to be a
number of samples, instead of a single dimension of a square of samples
(so an old option "-a3" must now be given as "-a9").