Rewrite SampleSet to allow sub-samples for each top-level sample
The number of sub-samples is specific to each sample channel.  The
number of sub-samples is stored in the SampleSet::Channel<T>::size
field.  By default, the number of sub-samples is 1, which makes the
interface largely compatible with the old one.

As part of this change, all samples are now stored in a contiguous
vector, instead of having a vector of samples for each channel; the
channel then just stores a offset into the appropriate vector (and
number of sub-samples).

The basic interface to SampleGen is now changed: SampleGen objects now
just accept a vector iterator to store the result into, and no longer
allocate their own vector for the result (so the caller now does any
necessary allocation of space); the SampleGen::adjust_sample_count
method allows the number of samples to be perturbed if the sample
generator would like to do so.

Because of presence of sub-samples, the number of samples can now vary
for different channel, so the SampleGen class is changed to not have
fixed number of samples (the number of top-level samples is now stored
in the SampleSet object in the SampleSet::num_samples field).