Move integrator and sample state into RenderContext
Instead of creating the surface integrator locally inside
Renderer::render_pixel, make it part of the RenderContext.

This makes it easily available anyplace with a reference to a
RenderContext.  It should also be a bit more efficient.

Also move the SampleSet state out of Renderer into RenderContext,
so it will be more easily available, and its lifetime more
obviously coupled with that of the integrator.