Remove SurfaceInteg::li method
The real interface exported by subclasses is SurfaceInteg::lo, and in
general it's probably better for callers to use that anyway (as the
caller probably also wants to call a volume integrator, and it's useful
to have the intermediate state resulting from the scene intersection
search to do that).

The only caller is currently Renderer::render_pixel, so just move the
code which used to be in SurfaceInteg::li into that method.