Remove surface projection term from BRDF evaluation functions
When calculating incoming light on a surface, a "cosine (theta_i)" term
is used to adjust for projected area on the surface (where theta_i is
the angle between the incoming light ray and the surface normal).

Previously this term was added as part of the BRDF evaluation functions,
but it's more common to define BRDFs without it.

So remove this term from the BRDFs and add it in the rendering core
instead.