Be more careful about calculation in Fresnel::reflectance
Avoid divide-by-zero errors by making sure the denominator can never be
zero, instead of explicitly checking for that case.

Use the identity cos^2+sin^2=1 to convert from sin to cos, to avoid the
high cost of trigonometric functions.

Clamp the result to the range 0-1, to avoid slightly straying outside
that range because of floating-point imprecision.