Avoid flipped eye-vector in CookTorranceBrdf::sample
Very rarely, the eye-vector is flipped (i.e., has a negative z value)
because of a discrepancy between the normal-frame and the
geometric-frame of the intersection.

In such cases, CookTorranceBrdf::sample used to return a negative pdf,
which caused odd results.

I'm not sure what the best way to handle such cases is, but for now,
just return "no sample" for these cases.