Add PathInteg::random_sample_set and PathInteg::random_direct_illum
These objects used to be local variables in PathInteg::Li, but they
took a non-trivial amount of time to initialize, allocate memory
upon initialization, and aren't even used for short paths.

Making them fields in the PathInteg object saves all this
unnecessary initialization time in PathInteg::Li, which is called
once per eye-ray.

It means we have to be slightly more careful about reentrancy
issues, but this is manageable, as a given PathInteg object should
only be used within a single thread, and the places where
PathInteg::Li might be called recursively are limited; specifically,
it should only happen during a call to VolumeInteg::Li, and such
recursive usage should be fine, as random_direct_illum and
random_sample_set are only actively in use when computing
direct-lighting.