Change implementation of Media stack in PathInteg::Li
We used to use a std::list object to hold the stack of Media objects in
PathInteg::Li, but that's unnecessarily inefficient.

Instead, just keep a pointer to the top of the stack, and allocate new
objects on push using the fast Mempool allocator in CONTEXT.  Media
objects already have a parent pointer which we can use to pop the stack.