Handle sample-count adjustment better in SampleSet
Keep track of the number returned by SampleGen::adjust_sample_count
for each channel, and pass that exact number back to
SampleGen::gen_samples.

This removes the requirement for the adjusted number to be an exact
multiple of the number of top-level samples (although when it's not,
some samples will end up being unused), and so avoids an assertion
failure when the user specifies parameters such that it can't be.