Handle null integrator global state in RenderContext constructor
In some cases we want to create a RenderContext object when the global
state isn't entirely setup yet, and in such cases, the GlobalRenderState
surface_integ or volume_integ fields can be a null pointer.

In such cases, just skip creating the corresponding integrator and give
it a null value, instead of crashing.