Add RecursiveInteg class
RecursiveInteg is an abstract superclass for surface-integrators
that use simple recursion to handle perfectly specular surfaces.
This class will handle the recursion for such surfaces and also any
direct emission by surfaces, and call a subclass-provided method to
handle other lighting.

Most of this code used to be in DirectInteg (which is now a
subclass of RecursiveInteg).