Restructure handling of environment maps
Instead of having a hard-wired environment map in the Scene object,
change the Scene::background method to just iterate over the scene's
list of environmental lights to compute background radiance.

This removes the all the various background-related scene fields and
methods, and is generally much cleaner.

This also makes other types of environmental lights properly
"visible" -- i.e., can be seen directly by the camera or via
specular reflection -- instead of just the environment map.

Some functionality is lost, in particular the ability to have a
separate light-map and environment-map, and the somewhat clever code
that used to be in Scene::background for adding HDR light-map
highlights to a LDR environment-map.  That functionality can later
be re-added in a cleaner way.