Make PhotonInteg final-gathering recurse for specular surfaces
When a final-gathering sample hits a specular surface, there's
nothing for it to gather, as photons are not deposited on purely
specular surfaces.

To help address this, recursively sample purely specular surfaces
during final-gathering, up to small fixed depth (randomly choosing
only a single sample direction for specular surfaces with more than
one).

This generally solves the problem, at the cost of greater rendering
time in scenes with many specular surfaces, and a small increase in
noise.