Filter parameters passed to surface-integrator constructors
When creating the surface-integrator in
GlobalRenderState::make_surface_integ_global_state, make a table
containing only parameters intended for the integrator, stripping off
the "surface-integ.NAME." prefix from the names of parameters, instead
of just passing in the raw global parameter table (which is available as
GlobalRenderState::params anyway).

This makes parsing in each integrator's constructor a bit more
convenient, although they now have to go to some extra trouble to get
parameters not specifically intended for them.

Also do some pruning of parameter names, removing aliases which are
unlikely to be used.