Rewrite EnvmapLight based on Hist2dPdf
Use Hist2dPdf with the environment map image as an input, instead of
the hairy (and somewhat low-quality) StructLight.  This allows us to
sample the environment-map much more accurately and more simply.

Also remove all the "dual sampling" stuff in EnvmapLight, which
sampled the hemisphere as well as the environment map; this is not
necessary now that we can do better intensity sampling.