Add pdf method variants in Dist subclasses, that take a Vec argument
That is, CosDist::pdf, WardDist::pdf, PhongDist::pdf, and
HemiDist::pdf (the latter just ignores the argument).

They're all symmetric distributions, and so the old pdf methods
(whose arguments were all the cosine of the angle with the normal),
still work fine, but a pdf method that takes the direction is more
general and it's nice if one can treat distributions in a somewhat
uniform manner.