Include an extra 1/NUM_SAMPLES factor in mis_sample_weight
In places where we scale by the result of mis_sample_weight, we
basically always want to subsequently divide by NUM_SAMPLES (since
we use mis_sample_weight when averaging samples).

Doing so implicitly in mis_sample_weight seems very natural, is a
bit more efficient (it avoids a divide), and allows callers to be a
bit simpler.

Note that this change does not affect any existing code, as
mis_sample_weight is currently only called with NUM_SAMPLES == 1.