Fix direction/position mapping in DirHist to be equal-area
To do this, we change DirHist::pos_to_dir and DirHist::dir_to_pos to
use a cylinder-style mapping, where the position V coordinate is
mapped directly to the _height_ of directional vectors (the Z
coordinate), rather than the angle from the XY plane as was used
previously.

Using this type of mapping, the "bunching up" of the position U
coordinate as directions approach the Z-axis is compensated for by
the "spreading out" of the V coordinate in the same situation, so
that the resulting area remains constant.

[The main use of DirHist is for sampling with DirHistDist, and for
this, it's important that all the histogram bins represent equal
area in the sample space.]