Tweak PhotonMap::find_photons interface
+ Replace the MAX_DIST argument with MAX_DIST_SQ.

This is slightly more annoying for callers in some cases, but it more
naturally represents what's going on, and better matches the form of the
return value (which is also max-distance-squared).

+ Always maintain the result vector in heap-form, even when there are
  fewer than MAX_PHOTONS elements, and rename the argument to
  PHOTON_HEAP.

This makes the code simpler, and makes it clear to callers what exactly
the contents of the result vector are.

These changes are intended to make it simpler to call
PhotonMap::find_photons repeatedly to build up a result vector from
multiple photon-maps.