Make ImageOutput buffering size-independent
Previously ImageOutput buffered a fixed number of output rows, the
number of which had to be declared in advance by the creator.

Instead of doing that, simply buffer as many rows as necessary, until
the caller declares that the earlier rows are finished (and so can be
written out) by calling ImageOutput::set_min_y.