Fix segfault in Mesh::Triangle::intersects
The code in Mesh::Triangle::intersects which tries to avoid funny
shadows for smoothed meshes uses Intersect::no_self_shadowing to get
the triangle which is potentially shadowed by the triangle being
shadow-tested.  This is hacky but works, because we know the two
objects lie within the same mesh (as Instance::smoothing_group
points to the mesh), and that no_self_shadowing is set by
Mesh::Triangle::IsecInfo::make_intersect.

However when a mesh is inside an instance this fails, because
Instance::IsecInfo::make_intersect clears the no_self_shadowing
field.

To avoid crashing in this case, just detect the null pointer and
disable the test in that case.