Add LocalXform class
This is a helper class for local/world tranformations, which is easy
to use as a second superclass of various other classes that operate in
a local coordinate system.

Use it as a superclass for the LocalSurface class (replacing most of
it), and the LocalPrimitive class, allowing it to inherit from
Primitive instead of LocalSurface, without duplicating any code.