Properly handle intensity-scaled output in image recovery
Because the ImageOutput class applies intensity-scaling
(intensity_scale, intensity_power) when writing the buffered output
image to the output file, we have to be careful during image recovery,
as the values read from the image being recovered have already had the
scaling applied.

To do this, the recover_image function temporarily turns off
intensity-scaling and when finished recovering, flushes the buffered
image to the output file before restoring the intensity-scaling values
to be used for rendering.