Add sample-offsets to ImageOutput
This adds two optional parameters, "sample-base-x" and "sample-base-y",
both of which can be specified using PARAMS in the constructor.

These x and y values are subtracted from sample coordinates passed to
ImageOutput::add_sample.  This is used to correctly handle "output
limits" (specified with the -L/--limit option) with a non-zero offset.

The method OutputImage::set_min_y is renamed to
OutputImage::set_sample_min_y, to emphasize the fact that its
y-coordinate argument is in the sample coordinate-system (as opposed
to OutputImage::set_raw_min_y, where the argument is a y-coordinate in
the 0,0-based output-image coordinate-system).