Don't prime negative intersect cache in Space::intersects
The effect of this priming is very subtle and hard to detect, and it's
the only real use of the ShadowRay class.  Not doing the priming seems
to have an extremely small effect, and should allow uses of ShadowRay
to be replaced by Ray (which is a more general class).