Replace sphere_intersect function with sphere_intersects
sphere_intersects is basically the same thing, but with a slightly
more consistent interface:  it always returns a boolean saying whether
an intersection occurred, and then returns the actual parametric
distance in an out-parameter.

The interface is also cleaned up a bit in other ways, such as always
taking a sphere center argument.

The MIN_T parameter is now checked more thoroughly, only returning
true if the intersection parametric distance T is greater than MIN_T
(previously it would only check MIN_T for the first intersection).
This means that callers no longer need to check against a minimum
distance.