Replace cylinder_intersect function with cylinder_intersects
cylinder_intersects is basically the same thing, but with a slightly
more consistent interface:  it always returns a boolean saying whether
an intersection occurred, and then returns the actual parametric
distance in an out-parameter.

The new function has two variants: one which takes a (bounded) Ray,
and one which takes an origin and a direction (and so is unbounded).