Fix Matrix4 row/column order confusion
Previously much of the Matrix4/Xform code was written such that matrix
rows and columns were swapped compared to common convention.  This was
done fairly consistently, so many things worked OK.  However, when
using matrix multiplication to compose transforms, the operand order
was the reverse of common convention (so that "A * B" had the effect
of transform A followed by transform B).

This commit swaps the row/column addressing, and fixes all the
confusion.  More explicit naming conventions are used to hopefully
make things a bit more clear (for instance, variables are called "row"
and "column" rather than simply "i" and "j").  The main visible effect
is that the operand order for matrix multiplication is now reversed
from previously (and is now consistent with convention):  Now "A * B"
has the effect of transform B followed by transform A.

Most uses of transform composition in the source tree have already
been switched to use functional syntax ("A (B)", meaning "transform B
followed by transform A") rather than multiplication, but this change
will cause any external uses of multiplication to compose
transformation matrices to be incorrect.