Tweak pretty-print syntax of Pos, Vec, Ray, and Xform classes
Switch to using { and } for bracketing (instead of < and >), and make
a few other tweaks.

The new syntax is slightly uglier, but has the advantage of being
compatible with Emacs calc (so one can just paste a printed Xform into
calc), Lua, and C.  Braces also interact better with Emacs structured-
movement commands.