snograw.swg: Add Lua-only Xform::operator(Ref<Tex<T>>) methods
This allows transforms to be applied to textures in Lua, returning a
new texture with a transform wrapper referencing the argument.

This is annoying to do in C++, but it's fairly easy to do with SWIG,
and makes more sense in Lua where the distinction between "by value"
and "by reference" is less rigid.