Add lua_read_file function
This is a C function intended for Lua programs to use, which reads the
entire contents of a file as a Lua string.  It tries to do so using
mmap, which is more efficient and less prone to memory problems than
using Lua functions to do the same thing.

When it's not able to read anything (e.g., because mmap isn't
supported), it just returns false so that the Lua caller can try using
Lua functions.