Make -s/--size option processing more generic
Instead of handling the -s/--size option separately in each
application, do most of the work in image-cmdline.{cc,h}.

Besides removing a bunch of code, this means snogcvt can now properly
handle a -s/--size option that only specifies a "size", not both width
and height, which is convenient.

It also stores the image size parameters in the image-parameters
valtable, which will allow scene loaders to examine and modify them.