Fix cone_sample distribution
The value of z needs to be proportional to the cosine of the cone's
apex half-angle, not proportional to the apex angle itself.

Just have the caller pass in the cosine directly (instead of
calculating it inside cone_sample), as it's quite often already
available to the caller anyway.

Also update all uses appropriately.