Get rid of russian-roulette continue cap in PathInteg
Previously the probability that russian-roulette would allow a path to
continue in PathInteg was capped at 0.5.  This meant that paths would
more often terminate early, but boosted the intensity of
non-terminated paths (after russian-roulette, the intensity is reduced
proportionally to the probability of continuation to keep the
calculation unbiased).

This cap is now removed, as many renders seem significantly less noisy
without it, with no appreciable increase in rendering time:  most
paths are still terminated fairly early by RR, but removing the cap
avoids "speckles" in the image that may have come from high-value
paths being terminated too early.