Don't use Tint multiply/divide operators for colors
The Tint class has operator* etc mostly so it can easily be used with
the FilterConv class, but they don't do the proper thing when the
intent is to actually scale the color value, as when adjusting image
intensity.

Change such cases to operate on the color explicitly instead of using
the Tint operators.