Make MempoolAlloc default constructor valid C++
MempoolAlloc has a default constructor, although using it would be
incorrect, because STL constructors want it (but never use it).
It refererences an undefined function to ensure linking fails if it
is used inadvertently.

However it should still be valid C++, so make sure it is (previously
it did not initialize the reference field MempoolAlloc::mempool).