Replace non-integral class constants with inline functions
C++ allows named constants in classes, in the form of fields declared
"static const" -- but annoyingly (and for no obvious reason), only if
they have an integral type.

g++ allows floating-point named class constants as an extension, which
is what snogray has been using up until now, but sadly, they are not
standard.

In the future, a standard-compatible form of such non-integral class
constants will be available via the C++0x "constexpr" feature, but
almost no compilers seem to really support it yet, and it will
probably be a long time before it's widely available.

This commit changes all the snogray constants to instead use inline
static methods, which are what is generally used in this case.
Hopefully these can be eventually replaced using constexpr.

The good news is that snogray now compiles without warning using g++
-Wall -Wextra -pedantic.  Snogray also compiles successfully using the
clang compiler.