Add hack to avoid double-define caused by <jpeglib.h>
jpeglib.h (or rather jconfig.h, which jpeglib.h includes) carelessly
defines HAVE_STDLIB_H, which conflicts with our definition of it in
"config.h".

This is normally harmless, but if -pedantic-errors or -Werror are
used, can cause the compiler to abort, so use an ugly hack to avoid
the conflict.