Clean up TupleMatrix interface a bit
Get rid of the BORDER argument which was passed around all over the
place, but rarely (never?) actually used.  Instead, make sure all
constructors and load and save methods have a PARAMS argument, and get
a "border" parameter from that instead.

PARAMS will also be useful for other purposes in the future.