configure.ac: Be more careful checking compiler options
Some compilers (e.g. clang, and some versions of gcc) don't consider
it a fatal error when they are given an unrecognized option, and
instead print a warning message on stderr and continue as normal.

Previously SNOGRAY_CHECK_CXX_FLAGS and SNOGRAY_CHECK_CXX_FLAG just
looked at compiler exit status, and so would come to the conclusion
that an option is supported even if the compiler gave a warning saying
it is not.

To handle this case, change those macros to look at the compiler error
output even when the exit status indicates success; if the option
names being tested occurs in the error output, then consider that a
failure message, and mark the option as unsupported.