Don't reference-count Bsdf objects
They are always allocated using a special arena and freed in bulk, so
reference-counting is not needed for memory management.

The only effect of reference-counting was to call Bsdf destructors at
some appropriate time, but really it's better to just make a rule that
Bsdf objects should have trivial destructors (as do all of the current
ones), and simply not call them at all.

Any future Bsdf subclasses that wish to allocate additional memory can
do so using the same special arena that they are allocated from, and
thus avoid any memory-management issues.