Add Mesh intersection methods
These are not normally called during rendering -- instead, individual
triangles in a Mesh are added to acceleration structures, and their
intersection methods are called instead -- but having them will keep
the Mesh class concrete when we make Surface an abstract class.

It's also nice to have them just for completeness (a user might, for
instance want to use them on a surface not yet added to any
acceleration structure).