Don't use pkg-config when cross-compiling
... as it often does the wrong thing.

This is true even with a recent version of the pkg-config
PKG_PROG_PKG_CONFIG autoconf macro, which uses AC_CHECK_TOOL to look
for a cross-version of pkg-config (${host}-pkg-config), because in the
case that a cross-version doesn't exist, it defaults to a non-cross
version, which is wrong when cross-compiling (the right thing would be
to just set PKG_CONFIG to "" in that case).

In the future this problem with PKG_PROG_PKG_CONFIG may be fixed, but
for now it's easiest to just not use when cross-compiling
(unfortunately losing the use of any libraries that depend on it).