Fix color alpha-unscaling in ImageOutput::set_raw_min_y
We can't use Tint::unscaled_color to get the alpha-unscaled color, as
the alpha value in the pixel will be wrong if it's been scaled by the
pixel's sampling weight.

Instead, just grab the raw alpha-scaled color, and then do the
alpha-unscaling ourselves after adjusting the alpha value by the
sampling weight if necessary.