Refactor Progress class
Make Progress an abstract class, and move the terminal-specific
progress-indicator implementation to a new TtyProgress subclass.

Also tweak the interface a bit, moving the setting of the progress
range from the constructor to the new methods Progress::set_start and
Progress::set_size.  This improves the abstraction, as the constructor
now only deals with target-specific initialization, and the actual
definition of the progress-range can be done elsewhere.  It also
leaves open the possibility of dynamically updating the progress
range.