Use FloatExceptsGuard to enable fp exceptions during rendering
Previously, fp exceptions were enabled in the main function after
loading was complete, and then never disabled.

Doing it only around rendering, and restoring the previous fp
exception state when rendering is done, will be safer when we get rid
of the strict separation of loading and rendering into sequential
phases (enabling fp exceptions during loading is problematic, because
loading uses many more external libraries, which are usually not
prepared for fp exceptions).