Correctly propagate Lua errors in luajit_exception_wrapper
When Lua code calls into C++, the C++ code then recursively calls some
Lua code, and an unhandled error occurs in the inferior Lua code, then
the resulting C++ exception raised by the inferior Lua is an opaque
type, not a string or subclass of std::exception (which
luajit_exception_wrapper already handled properly).

However in this case the actual Lua error value is still at the top of
the Lua stack, so when luajit_exception_wrapper catches an unknown
exception type, we just leave it in place and invoke lua_error (in the
superior Lua).

Of course if an unknown exception type wasn't actually raised by an
inferior Lua, this makes the resulting Lua error nonsense, but the
error in that case would be pretty useless regardless.  In the
usual case where it is from an inferior Lua, it's a huge win to
correctly propagate Lua errors.