Count translucent samples as "specular" in path integrator
Previously they were explicitly treated as non-specular (despite being
flagged as specular), with the comment:

   We "don't count" Bsdf::TRANSLUCENT for this purpose (even
   though they're nominally "specular"), because they actually
   _are_ accounted for in the direct-lighting term.

However, this seems to be simply wrong.  When the path integrator
samples direct lighting, it excludes Bsdf::SPECULAR from the sampling
flags (and the only place Bsdf::TRANSLUCENT is used is in the Stencil
class, and Stencil::sample only includes translucent samples if
Bsdf::SPECULAR is present).

Maybe in a previous incarnation of the code, the above comment was
correct, but it doesn't seem to be any longer.